Move remaining global types
