+++ dotcom-rendering/index.d.ts
@@ -1,91 +1,3 @@
-type SharePlatform =
-	| 'facebook'
-	| 'twitter'
-	| 'email'
-	| 'whatsApp'
-	| 'linkedIn'
-	| 'messenger';
-
-// shared type declarations
-
-interface SectionNielsenAPI {
-	name: string;
-	apiID: string;
-}
-
-interface Pagination {
-	currentPage: number;
-	totalPages: number;
-	newest?: string;
-	newer?: string;
-	oldest?: string;
-	older?: string;
-}
-
-type ContentType =
-	| 'article'
-	| 'network'
-	| 'section'
-	| 'imageContent'
-	| 'interactive'
-	| 'gallery'
-	| 'video'
-	| 'audio'
-	| 'liveBlog'
-	| 'tag'
-	| 'index'
-	| 'crossword'
-	| 'survey'
-	| 'signup'
-	| 'userid';
-
-type StageType = 'DEV' | 'CODE' | 'PROD';
-
-type CardImageType = 'picture' | 'avatar' | 'crossword' | 'slideshow' | 'video';
-
-type SmallHeadlineSize =
-	| 'tiny'
-	| 'small'
-	| 'medium'
-	| 'large'
-	| 'huge'
-	| 'ginormous';
-
-type MediaType = 'Video' | 'Audio' | 'Gallery';
-
-type LeftColSize = 'compact' | 'wide';
-
-type UserBadge = {
-	name: string;
-};
-
-// ------------
-// Liveblogs //
-// ------------
-type LiveUpdateType = {
-	numNewBlocks: number;
-	html: string;
-	mostRecentBlockId: string;
-};
-
-// ------------
-// RichLinks //
-// ------------
-type RichLinkCardType =
-	| 'special-report'
-	| 'live'
-	| 'dead'
-	| 'feature'
-	| 'editorial'
-	| 'comment'
-	| 'podcast'
-	| 'media'
-	| 'analysis'
-	| 'review'
-	| 'letters'
-	| 'external'
-	| 'news';
-
 // ------------------------------
 // 3rd party type declarations //
 // ------------------------------
