+++ dotcom-rendering/src/components/Pagination.amp.stories.tsx
@@ -1,3 +1,4 @@
+import type { PaginationType } from '../types/liveBlog';
 import { Pagination } from './Pagination.amp';
 
 export default {
@@ -5,7 +6,7 @@ export default {
 	title: 'AMP/Components/Pagination',
 };
 
-const pagination: Pagination = {
+const pagination: PaginationType = {
 	currentPage: 1,
 	totalPages: 2,
 	oldest: '?page=with:block-5b3ad778e4b05bcf5a69676d',
@@ -18,7 +19,7 @@ export const PaginationDefault = () => (
 	<Pagination pagination={pagination} guardianURL="" />
 );
 
-const paginationOlder: Pagination = {
+const paginationOlder: PaginationType = {
 	currentPage: 1,
 	totalPages: 2,
 	oldest: '?page=with:block-5b3ad778e4b05bcf5a69676d',
@@ -29,7 +30,7 @@ export const PaginationWithOlder = () => (
 	<Pagination pagination={paginationOlder} guardianURL="" />
 );
 
-const paginationNewer: Pagination = {
+const paginationNewer: PaginationType = {
 	currentPage: 1,
 	totalPages: 2,
 	newest: '?page=with:block-5b3ad778e4b05bcf5a69676d',
