+++ dotcom-rendering/src/components/Pagination.amp.tsx
@@ -5,6 +5,7 @@ import ChevronLeftDouble from '../static/icons/chevron-left-double.svg';
 import ChevronLeftSingle from '../static/icons/chevron-left-single.svg';
 import ChevronRightDouble from '../static/icons/chevron-right-double.svg';
 import ChevronRightSingle from '../static/icons/chevron-right-single.svg';
+import type { PaginationType } from '../types/liveBlog';
 
 const paginationStyle = css`
 	${textSans12};
@@ -46,7 +47,7 @@ const marginRightStyle = css`
 `;
 
 type Props = {
-	pagination?: Pagination;
+	pagination?: PaginationType;
 	guardianURL: string;
 };
