+++ dotcom-rendering/src/lib/sharing-urls.ts
@@ -1,3 +1,5 @@
+import type { SharePlatform } from '../types/layout';
+
 const appendParamsToBaseUrl: (
 	baseUrl: string,
 	params: {
