+++ dotcom-rendering/src/model/article-sections.ts
@@ -1,3 +1,8 @@
+interface SectionNielsenAPI {
+	name: string;
+	apiID: string;
+}
+
 const sections = [
 	{
 		name: 'Guardian',
