+++ dotcom-rendering/src/types/content.ts
@@ -4,6 +4,23 @@ export type StarRating = 0 | 1 | 2 | 3 | 4 | 5;
 
 export type BoostLevel = 'default' | 'boost' | 'megaboost' | 'gigaboost';
 
+export type ContentType =
+	| 'article'
+	| 'network'
+	| 'section'
+	| 'imageContent'
+	| 'interactive'
+	| 'gallery'
+	| 'video'
+	| 'audio'
+	| 'liveBlog'
+	| 'tag'
+	| 'index'
+	| 'crossword'
+	| 'survey'
+	| 'signup'
+	| 'userid';
+
 // -------------------------------------
 // Elements
