+++ dotcom-rendering/src/types/layout.ts
@@ -1,0 +1,41 @@
+export type SharePlatform =
+	| 'facebook'
+	| 'twitter'
+	| 'email'
+	| 'whatsApp'
+	| 'linkedIn'
+	| 'messenger';
+
+export type RichLinkCardType =
+	| 'special-report'
+	| 'live'
+	| 'dead'
+	| 'feature'
+	| 'editorial'
+	| 'comment'
+	| 'podcast'
+	| 'media'
+	| 'analysis'
+	| 'review'
+	| 'letters'
+	| 'external'
+	| 'news';
+
+export type CardImageType =
+	| 'picture'
+	| 'avatar'
+	| 'crossword'
+	| 'slideshow'
+	| 'video';
+
+export type SmallHeadlineSize =
+	| 'tiny'
+	| 'small'
+	| 'medium'
+	| 'large'
+	| 'huge'
+	| 'ginormous';
+
+export type MediaType = 'Video' | 'Audio' | 'Gallery';
+
+export type LeftColSize = 'compact' | 'wide';
