+++ dotcom-rendering/src/types/liveBlog.ts
@@ -1,0 +1,14 @@
+export type LiveUpdateType = {
+	numNewBlocks: number;
+	html: string;
+	mostRecentBlockId: string;
+};
+
+export interface PaginationType {
+	currentPage: number;
+	totalPages: number;
+	newest?: string;
+	newer?: string;
+	oldest?: string;
+	older?: string;
+}
